Added recipes and ingredients. Missing Create, Update, Delete for them.
Next step have to create the Create Update and Delete for Recipe and Ingredients. Then react UI. Might have to look up routing for that. On creation of a recipe or ingredient, might have to look at ways of updating RecipeIngredient table. Had a bit of trouble with creating that table sqllite. Turns out I was always looking at old unstopped DB.